% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.R
\name{congressTimeline}
\alias{congressTimeline}
\title{Takes list of twitter handles and last names for each session of congress.
Runs rtweet::get_timelines to access twitter api using twitter handles specified by user.}
\usage{
congressTimeline(x, y, n_tweets)
}
\description{
Takes list of twitter handles and last names for each session of congress.
Runs rtweet::get_timelines to access twitter api using twitter handles specified by user.
}
