% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvOut.R
\name{csvOut}
\alias{csvOut}
\title{Takes tweets from previous function and creates a csv for each member of congress where there are reported tweets.
Titles each csv as "ScrapeCongress i" where i represents the alphabetical order in which that member of Congress appears in the list that the user selected when entering in the timeline option in ScrapeCongress().}
\usage{
csvOut(tweets)
}
\description{
Takes tweets from previous function and creates a csv for each member of congress where there are reported tweets.
Titles each csv as "ScrapeCongress i" where i represents the alphabetical order in which that member of Congress appears in the list that the user selected when entering in the timeline option in ScrapeCongress().
}
