% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrapeCongress}
\alias{scrapeCongress}
\title{Takes tweets from previous function and creates a csv for each member of congress where there are reported tweets.
Titles each csv as "ScrapeCongress i" where i represents the alphabetical order in which that member of Congress appears in the list that the user selected when entering in the timeline option in ScrapeCongress().}
\usage{
scrapeCongress(
  timelines = "all",
  session = "116",
  n_tweets = 50,
  csvOut = FALSE
)
}
\description{
Takes tweets from previous function and creates a csv for each member of congress where there are reported tweets.
Titles each csv as "ScrapeCongress i" where i represents the alphabetical order in which that member of Congress appears in the list that the user selected when entering in the timeline option in ScrapeCongress().
}
